chore(json-crdt): 🤖 rename /con to /const folder as "con" is reserved path on Windows
