+++ src/json-crdt/nodes/const/ConNode.ts
